SDL: fix font size in windows
