+++ autogen.sh
@@ -1,10 +1,12 @@
 # This file is part of SmallBASIC
 #
-# Copyright(C) 2001-2012 Chris Warren-Smith.
+# Copyright(C) 2001-2015 Chris Warren-Smith.
 #
 # This program is distributed under the terms of the GPL v2.0 or later
 # Download the GNU Public License (GPL) from www.gnu.org
 #
+
+ln -sf README.md README
 
 pkg-config --version > /dev/null || echo "Please install pkg-config"
