+++ ide/android/assets/main.bas
@@ -56,7 +56,7 @@ sub do_about()
   print "(_ ._ _  _.|||_) /\ (_ |/ "
   print "__)| | |(_||||_)/--\__)|\_"
   print
-  print "Version 0.11.17"
+  print "Version 0.11.18"
   print
   print "Copyright (c) 2002-2015 Chris Warren-Smith"
   print "Copyright (c) 1999-2006 Nic Christopoulos" + chr(10)
